requirements.txt
numpy
scipy
torch
gymnasium
matplotlib
wandb
pyyaml
stable-baselines3
tqdm
ipython
